apps/ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
scikit-learn==1.4.0
numpy==1.26.3
pydantic==2.6.0
python-dotenv==1.0.0
